mcp/crash-course/4a-azure-openai-integration/requirements.txt
# Azure OpenAI MCP Integration Requirements

# Core Azure OpenAI and authentication
openai>=1.30.0
azure-identity>=1.15.0

# Model Context Protocol
mcp>=1.0.0

# Environment and async support
python-dotenv>=1.0.0
nest-asyncio>=1.5.0

# Additional utilities
asyncio-pool>=0.5.0
